Rearranged utils and refactored a little bit of the index
