+++ lib/api/giantbomb.js
