+++ lib/imageblur.js
