+++ routes/backend.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 
 const env = require('dotenv').config()
-const gb = require('../lib/queryGB')
+const gb = require('../lib/api/giantbomb')
 
 const moment = require('moment')
